// types/ProductItem.ts
export type ProductItem = {
  backbox_grade_label: 'Excellent' | 'Good' | 'Fair';
  backmarketID: number;
  brand: string;
  brand_label: string;
  categories: { lvl0: string; lvl1: string };
  category_1: string;
  category_2: string;
  category_3: string;
  color: string;
  color_hex: string;
  compatible_carriers: Array<{ label: string; logo: string }>;
  currency: string;
  date_release: string;
  image1: string;
  is_price_sexy: boolean;
  link_grade_v2: { href: string };
  nb_sales_model: number;
  nb_sales_model_log10: number;
  price: number;
  price_new: number;
  review_stars: number;
  shipping_delay: string[];
  storage: string;
  storage_label: string;
  sub_title_elements: string[];
  title: string;
  title_model: string;
  warranty: number;
};